+ Fast I/O dispatch is now displayed (you need to explicitly instruct the driver to collect it) (the collection may be dangerous)  + Information related to devnode tree may be collected by the driver if you wish so (however, this may be dangerous)  + addresses of some routines are translated to their names (IopInvalidDeviceRequest and several from the Fast I/O list)  + new type of device tree view: PnP tree (very similar to the view you get when you instruct Device Manager to display devices by connection)  + device cescription and friendly name are now taken from resources if the original strings (obtained via IoGetDeviceProperty) advise so  + hardware and compatible IDs are now retrieved through IoGetDevicePropertyData. The IRP_MN_QUERY_ID is no longer used  + the GUI now displays a list of loaded device drivers. The following details are gathered    + image base address    + image sizeof    + file name    + signature state (not signed, signed, expired, timestamp expired, revoked)    + certificate Common Names (CN)  You can configure which signature checks you wish to perform (e.g. whether the time stamp certs are always trusted (even when they expire) which seems to be a standard Windows behavior (expired timestamp cert seems to mean nothing for the kernel when it comes to load a driver))
Well, this is a base for the v1.6 release. The biggest problem is that most of the new information does not appear in the log files generated by the tool.

When v1.6 release time comes, the driver should still be signed by a SHA-1 certificate so it should work well even on older systems. Actually, I signed it few minutes ago and I consider is a v1.6-ready. The reason for such a behavior is that SHA1 certs seem not to be trusted by some systems when used after January 1th.